--- src/apis/index.ts
@@ -1,2 +1,0 @@
-export * from './onMessage'
-export * from './sendMessage'
