+++ src/types.ts
@@ -1,8 +1,8 @@
-import { Jsonify, JsonValue } from 'type-fest'
+import type { JsonValue, Jsonify } from 'type-fest'
 
 export type RuntimeContext = 'devtools' | 'background' | 'popup' | 'options' | 'content-script' | 'window'
 
-export type Endpoint = {
+export interface Endpoint {
   context: RuntimeContext
   tabId: number
   frameId?: number
@@ -29,18 +29,13 @@ export interface IInternalMessage {
   timestamp: number
 }
 
-export interface IQueuedMessage {
-  resolvedDestination: string
-  message: IInternalMessage
-}
-
-export type StreamInfo = {
+export interface StreamInfo {
   streamId: string
   channel: string
   endpoint: Endpoint
 }
 
-export type HybridUnsubscriber = {
+export interface HybridUnsubscriber {
   (): void
   dispose: () => void
   close: () => void
@@ -70,8 +65,8 @@ export interface ProtocolMap {
 export type DataTypeKey = keyof ProtocolMap
 
 export type GetDataType<
-  K extends DataTypeKey | string,
-  Fallback extends JsonValue
+  K extends DataTypeKey | string = string,
+  Fallback extends JsonValue = undefined,
 > = K extends DataTypeKey
   ? ProtocolMap[K] extends ProtocolWithReturn<infer Data, any>
     ? Data
@@ -79,8 +74,8 @@ export type GetDataType<
   : Fallback
 
 export type GetReturnType<
-  K extends DataTypeKey | string,
-  Fallback extends JsonValue
+  K extends DataTypeKey | string = string,
+  Fallback extends JsonValue = undefined,
 > = K extends DataTypeKey
   ? ProtocolMap[K] extends ProtocolWithReturn<any, infer Return>
     ? Return
